Updated 02:32 1/11/2023

+++ app/src/main/java/com/example/qurantutor/data/ResponseData.kt
@@ -3,5 +3,5 @@ package com.example.qurantutor.data
 data class ResponseData(
     val language: String,
     val transcribed_text: String,
-    val bleu_score: Int
+    val bleu_score: Float
 )
